build: Added rest server demo application and made changes to pom
Wrote the neccesary code and made the necessary files to have an demo application. It is able to run a server and use GET. Instructions is inside DemoApplication.java. Added the necessary plugins and dependencies so we are able to run the RESTful server using springboot. Made also some minor changes to parent pom.

#54