feat(rest): Added serializers and deserializers
Made UnitListModule.java that conatins serializers and deserializers for Unit.java and UnitList.java. \n There reason for this is so that in DemoApplication.java we can use UnitListModule.java as bean and as such be able serialize/deserialize from Unit/UnitList objects. \n With this change we can now, for example, use Object Unit as RequestBody in POST.

#56