feat(rest): Made PUT-methode that removes tenant by location rather than name.
Made PUT-methode that removes tenant by location rather than name. The reason for this change is that a person can be registered as a tenant of several tenant. Therfore, using name to remove a tenant from a unit would cause issues if a tenant is registered with several units. As such, location is better to use.

#59